+++ R/Alakazam.R
@@ -147,6 +147,7 @@
 #'                          stri_pad_left stri_pad_right
 #'                          stri_detect_fixed stri_paste
 #' @importFrom	ape 		read.tree di2multi reorder.phylo root
+#' @importFrom	reshape2 	melt
 #' @importFrom  tidyr       complete_
 #' @importFrom  Rcpp        evalCpp
 #' @useDynLib   alakazam, .registration=TRUE